% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GetExpression}
\alias{GetExpression}
\title{Get Expression}
\usage{
GetExpression(expr.file, gene, resids = TRUE)
}
\arguments{
\item{expr.file}{file containing expression and interaction values}

\item{gene}{gene of interest}

\item{resids}{default true, plotting residuals after regressing out covariates}
}
\description{
Get Expression
}
