% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GetGenePosition}
\alias{GetGenePosition}
\title{Get gene starting and ending positions}
\usage{
GetGenePosition(gene.name, gene.info.file)
}
\arguments{
\item{gene.name}{gene name for which we want the Manhattan}

\item{gene.info.file}{hdf5 file containing gene information}
}
\description{
add gene start and gene end to dataframe for Manhattan plots
}
