% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GetGenotypes}
\alias{GetGenotypes}
\title{Get genotypes}
\usage{
GetGenotypes(geno.file.prefix, chrom, snp)
}
\arguments{
\item{geno.file.prefix}{prefix of hdf5 file containing genotype information}

\item{chrom}{chromosome of interest}

\item{snp}{SNP of interest}
}
\description{
Get genotypes
}
