% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GetInteractionFactor}
\alias{GetInteractionFactor}
\title{Get Interaction Term}
\usage{
GetInteractionFactor(interaction.file)
}
\arguments{
\item{interaction.file}{file containing expression and interaction values}
}
\description{
Get Interaction Term
}
