% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_pipeline_postprocessing.R
\name{GetPerms}
\alias{GetPerms}
\title{Get permutations from qtl pipeline for plotting purposes}
\usage{
GetPerms(results.folder)
}
\arguments{
\item{results.folder}{full path to folder where results are stored}
}
\description{
Get permutations from qtl pipeline for plotting purposes
}
