% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_pipeline_postprocessing.R
\name{GetResults}
\alias{GetResults}
\title{Get results from qtl pipeline and summarize results}
\usage{
GetResults(results.folder, multiple.testing.global = "ST")
}
\arguments{
\item{results.folder}{full path to folder where results are stored}

\item{multiple.testing.global}{method, can be BF (Bonferroni) or ST (Storey)}
}
\description{
All genes, one scan
}
