% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_pipeline_postprocessing.R
\name{GetResultsWithoutCorrection}
\alias{GetResultsWithoutCorrection}
\title{Get results from qtl pipeline without correction}
\usage{
GetResultsWithoutCorrection(results.folder)
}
\arguments{
\item{results.folder}{full path to folder where results are stored}
}
\description{
Get results from qtl pipeline without correction
}
