% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{MakeInteractionDataFrame}
\alias{MakeInteractionDataFrame}
\title{Make dataframe for Interaction plots}
\usage{
MakeInteractionDataFrame(interaction.file, geno.file.prefix, gene, snp,
  resids = TRUE)
}
\arguments{
\item{interaction.file}{interaction file containing int factor and expression}

\item{geno.file.prefix}{genotype file, per chromosome}

\item{gene}{gene of interest}

\item{snp}{SNP of interest}

\item{resids}{default true, plotting residuals after regressing out covariates}
}
\description{
Make dataframe for Interaction plots
}
