% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{MakeManhattanDataFrame}
\alias{MakeManhattanDataFrame}
\title{Make dataframe for Manhattan plots}
\usage{
MakeManhattanDataFrame(results.df, gene.name, gene.info.file)
}
\arguments{
\item{results.df}{dataframe containing results for all genes}

\item{gene.name}{gene name for which we want the Manhattan}

\item{gene.info.file}{hdf5 file containing gene information}
}
\description{
Make dataframe for Manhattan plots
}
