% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_pipeline_postprocessing.R
\name{MergeGeneInfo}
\alias{MergeGeneInfo}
\title{Merge results with gene info & snp info}
\usage{
MergeGeneInfo(results.folder, gene.names.df, snp.info.df,
  multiple.testing.global = "ST")
}
\arguments{
\item{results.folder}{full path to folder where results are stored}

\item{gene.names.df}{dataframe containing snp info, e.g. rsID}

\item{snp.info.df}{dataframe containing snp info, e.g. rsID}

\item{multiple.testing.global}{method, can be BF (Bonferroni) or ST (Storey)}
}
\description{
Merge results with gene info & snp info
}
