% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{OneManhattan}
\alias{OneManhattan}
\title{Generates the Manhattan plot for one set of results}
\usage{
OneManhattan(results.df, gene.name, gene.info.file, snp.name = "")
}
\arguments{
\item{results.df}{dataframe containing results, needs SNP p-values and positions}

\item{gene.name}{gene name for which we want the Manhattan}

\item{gene.info.file}{hdf5 file containing gene information}

\item{snp.name}{SNP of interest within the Manhattan plot (optional)}
}
\description{
One gene, one type of analysis
}
