% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotInteraction}
\alias{PlotInteraction}
\title{Generates an interaction plot}
\usage{
PlotInteraction(interaction.file, geno.file.prefix, gene, snp)
}
\arguments{
\item{interaction.file}{interaction file containing interaction factor & gene expression}

\item{geno.file.prefix}{genotype file, per chromosome}

\item{gene}{gene of interest}

\item{snp}{SNP of interest}
}
\description{
One gene, one SNP
}
