% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotQQ}
\alias{PlotQQ}
\title{Generates the qq plot for one set of results}
\usage{
PlotQQ(results.df, col = c("cornflowerblue", "black"))
}
\arguments{
\item{results.df}{dataframe containing results, needs SNP p-values and permutations}
}
\description{
All genes, one type of analysis
}
