% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{TwoManhattan}
\alias{TwoManhattan}
\title{Generates overlapping Manhattan plots for two or more sets of results}
\usage{
TwoManhattan(results.dfs, gene.name, gene.info.file, snp.name = "")
}
\arguments{
\item{results.dfs}{list of dataframes containing results, needs SNP p-values and positions}

\item{gene.name}{gene name for which we want the Manhattan}

\item{gene.info.file}{hdf5 file containing gene information}

\item{snp.name}{SNP of interest within the Manhattan plot (optional)}
}
\description{
One gene, comparison between two or more types of analysis
}
