% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_pipeline_postprocessing.R
\name{LeadSnpsOnly}
\alias{LeadSnpsOnly}
\title{Only save top lead per gene - optionally only significant}
\usage{
LeadSnpsOnly(results.df, significance.threshold = 1)
}
\arguments{
\item{results.df}{dataframe containing results}

\item{significance.threshold}{at global level (optional)}
}
\description{
Only save top lead per gene - optionally only significant
}
